app_producer/requirements.txt
python-dotenv
influxdb_client
paho-mqtt
MacTmp
psutil
mysql-connector-python 
Flask-MQTT
Flask
pika
Flask-Script
gunicorn